Added functionality directories
