+++ sail-app/app/src/main/java/com/functrco/sail/screens/main/cart/CartFragment.kt
@@ -1,18 +1,18 @@
-package com.functrco.sail.screens.main
+package com.functrco.sail.screens.main.cart
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import com.functrco.sail.R
-class SearchFragment : Fragment() {
 
+class CartFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_search, container, false)
+        return inflater.inflate(R.layout.fragment_cart, container, false)
     }
 }
