+++ sail-app/app/src/main/java/com/functrco/sail/screens/main/orders/OrdersFragment.kt
@@ -1,4 +1,4 @@
-package com.functrco.sail.screens.main
+package com.functrco.sail.screens.main.orders
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,12 +7,12 @@ import android.view.View
 import android.view.ViewGroup
 import com.functrco.sail.R
 
-class SettingsFragment : Fragment() {
+class OrdersFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_settings, container, false)
+        return inflater.inflate(R.layout.fragment_orders, container, false)
     }
 }
