+++ sail-app/app/src/main/res/layout/fragment_cart.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".screens.main.CartFragment">
+    tools:context=".screens.main.cart.CartFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
