+++ sail-app/app/src/main/res/layout/fragment_orders.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".screens.main.OrdersFragment">
+    tools:context=".screens.main.orders.OrdersFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
