+++ sail-app/app/src/main/res/navigation/nav_bottom.xml
@@ -13,23 +13,23 @@
 
     <fragment
         android:id="@+id/searchFragment"
-        android:name="com.functrco.sail.screens.main.SearchFragment"
+        android:name="com.functrco.sail.screens.main.search.SearchFragment"
         android:label="Search"
         tools:layout="@layout/fragment_search" />
 
     <fragment
         android:id="@+id/cartFragment"
-        android:name="com.functrco.sail.screens.main.CartFragment"
+        android:name="com.functrco.sail.screens.main.cart.CartFragment"
         android:label="Cart"
         tools:layout="@layout/fragment_cart" />
     <fragment
         android:id="@+id/ordersFragment"
-        android:name="com.functrco.sail.screens.main.OrdersFragment"
+        android:name="com.functrco.sail.screens.main.orders.OrdersFragment"
         android:label="Orders"
         tools:layout="@layout/fragment_orders" />
     <fragment
         android:id="@+id/settingsFragment"
-        android:name="com.functrco.sail.screens.main.SettingsFragment"
+        android:name="com.functrco.sail.screens.main.settings.SettingsFragment"
         android:label="Settings"
         tools:layout="@layout/fragment_settings" />
 </navigation>
